create json_response module under services
