+++ app/services/github_service_repo_search.rb
@@ -1,4 +1,5 @@
 class GithubServiceRepoSearch
+  include JsonResponse
 
   def initialize(user)
     @user = user
@@ -11,18 +12,5 @@ class GithubServiceRepoSearch
   private
 
     attr_reader :user
-
-    def conn
-      Faraday.new(url: "https://api.github.com", headers: headers)
-    end
-
-    def headers
-      {"access_token" => user.oauth_token}
-    end
-
-    def json_response(url)
-      response = conn.get(url)
-      JSON.parse(response.body, symbolize_names: true)
-    end
 
 end
